graph agent/requirements.txt
# BlockStat Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

networkx==3.2.1
leidenalg==0.10.1
python-igraph==0.10.8 

numpy==1.24.3
pandas==2.0.3

requests==2.31.0
httpx==0.25.0

python-dotenv==1.0.0
python-multipart==0.0.6

neo4j==5.14.0
sqlalchemy>=2.0

agno
groq
